fix: Убрать зависимость гранаты от того, находится ли игрок за стеной или нет, т.к. пока что не работает

+++ Source/LestaStart/Core/HPRenderComponent.cpp
@@ -27,7 +27,7 @@ FString UHPRenderComponent::MakeActualStringFromTemplate(const float NewHP) cons
 void UHPRenderComponent::OnHealthChanged(float NewHP)
 {
 	
-	SetText(FText::FromString(MakeActualStringFromTemplate(NewHP)));
+	SetText(FText::FromString(MakeActualStringFromTemplate(static_cast<int>(NewHP))));
 }
